Destroy sign_out route and add sign_out button to navigation panel
