+++ src/App.js
@@ -14,7 +14,6 @@ import Doctor from './components/Doctor';
 import './App.css';
 import SignUp from './components/auth_pages/Signup';
 import Login from './components/auth_pages/Login';
-import SignOut from './components/auth_pages/Signout';
 
 const router = createBrowserRouter(
   createRoutesFromElements(
@@ -27,7 +26,6 @@ const router = createBrowserRouter(
       <Route path="/doctors/:doctorId" element={<Doctor />} />
       <Route path="/sign_up" element={<SignUp />} />
       <Route path="/log_in" element={<Login />} />
-      <Route path="/sign_out" element={<SignOut />} />
     </>,
   ),
 );
