+++ src/components/auth_pages/Login.js
@@ -17,7 +17,7 @@ function LogIn() {
   const handleSubmit = async (e) => {
     e.preventDefault();
     try {
-      const response = await fetch('http://localhost:3000/users/sign_in', {
+      const response = await fetch('https://medimate-backend-p22y.onrender.com/users/sign_in', {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
